# public/assets/api.yaml
openapi: 3.0.0
info:
  title: 'My First API'
  version: '0.1'
paths:
...
  /userProfiles:
    get:
      tags:
        - UserProfile
      summary: 'Find list UserProfile'
      description: 'Returns list of UserProfile'
      operationId: getUserProfile
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/UserProfile' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/UserProfile' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/UserProfile' } }
                type: object
        '404':
          description: 'UserProfile not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - UserProfile
      summary: 'Add a new UserProfile to the store'
      operationId: addUserProfile
      requestBody:
        $ref: '#/components/requestBodies/UserProfile'
      responses:
        '201':
          description: 'Created UserProfile'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
            application/xml:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/userProfiles/{id}':
    get:
      tags:
        - UserProfile
      summary: 'Find UserProfile by ID'
      description: 'Returns a single UserProfile'
      operationId: getUserProfileById
      parameters:
        -
          name: id
          in: path
          description: 'ID of UserProfile to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
            application/xml:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'UserProfile not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - UserProfile
      summary: 'Update an existing UserProfile'
      operationId: updateUserProfile
      parameters:
        -
          name: id
          in: path
          description: 'UserProfile id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/UserProfile'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'UserProfile not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - UserProfile
      summary: 'Deletes a UserProfile'
      operationId: deleteUserProfile
      parameters:
        -
          name: id
          in: path
          description: 'UserProfile id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
  /userSocialites:
    get:
      tags:
        - UserSocialite
      summary: 'Find list UserSocialite'
      description: 'Returns list of UserSocialite'
      operationId: getUserSocialite
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/UserSocialite' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/UserSocialite' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/UserSocialite' } }
                type: object
        '404':
          description: 'UserSocialite not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - UserSocialite
      summary: 'Add a new UserSocialite to the store'
      operationId: addUserSocialite
      requestBody:
        $ref: '#/components/requestBodies/UserSocialite'
      responses:
        '201':
          description: 'Created UserSocialite'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSocialite'
            application/xml:
              schema:
                $ref: '#/components/schemas/UserSocialite'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/userSocialites/{id}':
    get:
      tags:
        - UserSocialite
      summary: 'Find UserSocialite by ID'
      description: 'Returns a single UserSocialite'
      operationId: getUserSocialiteById
      parameters:
        -
          name: id
          in: path
          description: 'ID of UserSocialite to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSocialite'
            application/xml:
              schema:
                $ref: '#/components/schemas/UserSocialite'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'UserSocialite not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - UserSocialite
      summary: 'Update an existing UserSocialite'
      operationId: updateUserSocialite
      parameters:
        -
          name: id
          in: path
          description: 'UserSocialite id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/UserSocialite'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'UserSocialite not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - UserSocialite
      summary: 'Deletes a UserSocialite'
      operationId: deleteUserSocialite
      parameters:
        -
          name: id
          in: path
          description: 'UserSocialite id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
...
components:
  schemas:
...
    UserProfile:
      title: UserProfile
      description: UserProfile
      properties:
        id:
          title: id
          description: id
          type: integer
          format: '-'
          maxLength: 10
          nullable: false
        user_id:
          title: user_id
          description: user_id
          type: integer
          format: '-'
          maxLength: 20
          nullable: false
        full_name:
          title: full_name
          description: full_name
          type: string
          format: '-'
          maxLength: 255
          nullable: false
        gender:
          title: gender
          description: gender
          type: string
          format: '-'
          maxLength: 255
          nullable: false
        birthday:
          title: birthday
          description: birthday
          type: string
          format: date
          nullable: false
        avatar:
          title: avatar
          description: avatar
          type: string
          format: '-'
          maxLength: 255
          nullable: false
        created_at:
          title: created_at
          description: created_at
          type: string
          format: '-'
          nullable: true
        updated_at:
          title: updated_at
          description: updated_at
          type: string
          format: '-'
          nullable: true
      type: object
    UserSocialite:
      title: UserSocialite
      description: UserSocialite
      properties:
        id:
          title: id
          description: id
          type: integer
          format: '-'
          maxLength: 20
          nullable: false
        user_id:
          title: user_id
          description: user_id
          type: integer
          format: '-'
          maxLength: 20
          nullable: false
        site:
          title: site
          description: site
          type: string
          format: '-'
          maxLength: 255
          nullable: false
        account_id:
          title: account_id
          description: account_id
          type: string
          format: '-'
          maxLength: 255
          nullable: false
        created_at:
          title: created_at
          description: created_at
          type: string
          format: '-'
          nullable: true
        updated_at:
          title: updated_at
          description: updated_at
          type: string
          format: '-'
          nullable: true
      type: object
...
  requestBodies:
...
    UserProfile:
      description: 'UserProfile object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/UserProfile'
        application/xml:
          schema:
            $ref: '#/components/schemas/UserProfile'
        application/json:
          schema:
            $ref: '#/components/schemas/UserProfile'
    UserSocialite:
      description: 'UserSocialite object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/UserSocialite'
        application/xml:
          schema:
            $ref: '#/components/schemas/UserSocialite'
        application/json:
          schema:
            $ref: '#/components/schemas/UserSocialite'
...